requirements.txt
fastapi
uvicorn
gradio_client
sqlite3
python-dotenv
langchain-google-genai